refactor ProfileViewController ProfileViewController for UI ProfileViewPresenter for logic
